To hell with stupid conversions.

+++ src/main/java/engsoft/progression/Arithmetic.java
@@ -1,14 +1,14 @@
 package engsoft.progression;
 
 // How I wish I had lambdas and currying ....
-public class Arithmetic implements Function1<Double, Double> {
-    private double base;
+public class Arithmetic implements Function1<Integer, Integer> {
+    private Integer base;
 
-    public Arithmetic(Double base) {
+    public Arithmetic(Integer base) {
 	this.base = base;
     }
 
-    public Double apply(Double previousValue) {
+    public Integer apply(Integer previousValue) {
 	return previousValue + base;
     }
 }
